FlaskServer/requirements.txt
torchvision
Flask
torch
torchvision
Pillow
python-dotenv